`parse_pac_file` now raises IOError
If the file containing pac information is not readable or does not
exist the current way of handling this is inadiquate (the exception
caught in the function and a print message sent to stdout via `print`).

This could and should be improved massively by sending back to the user
an IOError which can be handled in a sensible manner.