pymod: don't mistake embedded URLs for real URLs
Without this change, when passed an URL like

    http://www.google.com?q=http://www.yahoo.com

the python library would have mistakenly passed "www.yahoo.com" as the host
argument of the FindProxyForURL function.

Credit for finding and diagnosing the issue goes to lucacittadini@google.com

While at it, rename 'url_regex' => 'URL_REGEX', to honor the convention that
module-level constants should be in upper-case, improve a couple of error
messages, and remove a couple of extra empty lines.

Signed-off-by: Stefano Lattarini <slattarini@google.com>