Extend shell-based tests for pactester; not yet passing
This is part of the backporting of Google-specific modifications to
upstream pacparser.

Notice that the tests do not pass yet for the pactester build from the
git repository, but works from the one built from the Google-modified
version. My plan is to use the enhanced test to do the backporting in
a "test-driven" way.

Signed-off-by: Stefano Lattarini <slattarini@google.com>